# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Architecture Knowledge Platform API
  version: 0.2.0
  x-akp-write-idempotency: >
    Mutating POST operations accept Idempotency-Key and MCP write tools require
    it. POST /v1/auth/session is intentionally exempt because it creates a
    Set-Cookie response that cannot be safely replayed.
  description: >
    HTTP contract for the routes registered by the current Fastify runtime.
    All /v1 operations require a scoped bearer API token or an unexpired
    HttpOnly web session and enforce the permission recorded in
    x-akp-permission. Session-authenticated writes also require X-CSRF-Token.
servers:
  - url: http://127.0.0.1:8080
    description: Local development default
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /v1/ingest:
    get:
      tags: [Ingestion]
      operationId: listIngestJobs
      x-akp-permission: source:read
      responses:
        "200": { $ref: "#/components/responses/ObjectResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Ingestion]
      operationId: submitIngest
      x-akp-permission: source:write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestRequest"
      responses:
        "202":
          description: Job accepted or idempotent prior response returned
          content:
            application/json:
              schema:
                type: object
                required: [jobId, state]
                properties:
                  jobId: { type: string, format: uuid }
                  state: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/ingest/{id}:
    get:
      tags: [Ingestion]
      operationId: getIngestStatus
      x-akp-permission: source:read
      parameters:
        - { $ref: "#/components/parameters/UuidId" }
      responses:
        "200": { $ref: "#/components/responses/ObjectResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/ingest/{id}/cancel:
    post:
      tags: [Ingestion]
      operationId: cancelIngest
      x-akp-permission: source:write
      parameters:
        - { $ref: "#/components/parameters/UuidId" }
      responses:
        "200": { $ref: "#/components/responses/ObjectResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/ingest/{id}/retry:
    post:
      tags: [Ingestion]
      operationId: retryIngest
      x-akp-permission: source:write
      parameters:
        - { $ref: "#/components/parameters/UuidId" }
      responses:
        "200": { $ref: "#/components/responses/ObjectResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API token
      description: >
        Provisioned out of band with an explicit TokenScopeSet. Token scopes
        are intersected with the user's current memberships on every request.
    cookieAuth:
      type: apiKey
      in: cookie
      name: akp_session
      description: HttpOnly session carrying the effective TokenScopeSet snapshot.
  parameters:
...
    UuidId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    ObjectResponse:
      description: Runtime response object
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, expired, revoked or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Permission, space or path scope denied
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found in the authorized scope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Resource state, publication lock or optimistic revision conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [code]
      properties:
        code: { type: string }
      additionalProperties: true
...
    IngestRequest:
      type: object
      required: [spaceId, vaultId, sourceUri]
      properties:
        spaceId: { type: string, format: uuid }
        vaultId: { type: string, format: uuid }
        sourceUri: { type: string, minLength: 1 }
        expectedSha256: { type: string, pattern: "^[a-f0-9]{64}$" }
        title: { type: string }
        mediaType: { type: string }
        policy:
          type: string
          enum: [REVIEW_REQUIRED, ALLOW_LOW_RISK_AUTO_APPROVAL]
          default: REVIEW_REQUIRED
        idempotencyKey: { type: string, minLength: 8, maxLength: 200 }
...